refactor: use same root folder for example migration data
