+++ test/migrate/wordpressMigrationProvider.test.ts
@@ -5,7 +5,7 @@ import axios from 'axios';
 import * as fs from "fs-extra";
 
 import {WordpressMigrationProvider} from "../../src/migrate/wordpressMigrationProvider";
-import {posts, media, categories, tags} from "../wordpress_example/_mocks_";
+import {posts, media, categories, tags} from "../migration_example/wordpress/_mocks_";
 
 describe('wordpressMigrationProvider', () => {
