+++ test/migration_example/ghost/_mocks_.ts
