+++ test/migration_example/ghost/posts.ts
