+++ test/migration_example/wordpress/categories.ts
